+++ CHANGELOG.md
@@ -7,3 +7,4 @@ Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how 
 ## [Unreleased]
 
 - Initial release
+- Changed API key configuration method from settings.json to SecretStorage and Command Palette.